FixBug: GPS reading,Sound level Reading

+++ src/MQUnifiedsensor.h
@@ -50,7 +50,7 @@ class MQUnifiedsensor
     byte _firstFlag = false;
     float _VOLT_RESOLUTION  = 5.0; // if 3.3v use 3.3
     float _RL = 10; //Value in KiloOhms
-    byte _ADC_Bit_Resolution = 10;
+    float _ADC_Bit_Resolution = 10;
     byte _regressionMethod = 1; // 1 -> Exponential || 2 -> Linear
     
     float _adc, _a, _b, _sensor_volt;
